requirements.txt
datasets==2.16.1
effdet==0.4.1
ipykernel==6.28.0
loguru==0.7.2
openai==1.7.2
pickleshare==0.7.5
pikepdf==8.11.2
pip-chill==1.0.3
psycopg2-binary==2.9.9
pypdf==3.17.4
pytesseract==0.3.10
python-dotenv==1.0.0
sentencepiece==0.1.99
sqlalchemy==2.0.25
unstructured==0.12.0
unstructured-inference==0.7.21
unstructured.pytesseract==0.3.12
